Remove useActiveToolIdSynchedItems
